reuse locale code to enable url overriding
