+++ src/index.tsx
@@ -22,6 +22,7 @@ import { msalApplication } from './app/services/graph-client/msal-agent';
 import { DEFAULT_USER_SCOPES } from './app/services/graph-constants';
 import App from './app/views/App';
 import { readHistoryData } from './app/views/sidebar/history/history-utils';
+import { geLocale } from './appLocale';
 import messages from './messages';
 import { store } from './store';
 import './styles/index.scss';
@@ -75,29 +76,6 @@ msalApplication.acquireTokenSilent({ scopes: DEFAULT_USER_SCOPES.split(' ') }).t
 }).catch(() => {
   // ignore the error as it means that a User login is required
 });
-
-const localeMap: any = {
-  'de-de': 'de-DE',
-  'en-us': 'en-US',
-  'es-es': 'es-ES',
-  'fr-fr': 'fr-FR',
-  'ja-jp': 'ja-JP',
-  'pt-br': 'pt-BR',
-  'ru-ru': 'ru-RU',
-  'zh-cn': 'zh-CN'
-};
-
-function getTryItLocale() {
-  return new URLSearchParams(location.search).get('locale');
-}
-
-function getPortalLocale(): string {
-  return location.pathname.split('/')[1].toLocaleLowerCase();
-}
-
-const hostDocumentLocale = getTryItLocale() || getPortalLocale();
-
-const geLocale = hostDocumentLocale && localeMap[hostDocumentLocale] || 'en-US';
 
 addLocaleData([
   ...br,
